#809: Applied patch by Thomas Beutlich to fix Windows ccompilation error.

+++ trunk/src/nikonmn.cpp
@@ -2478,7 +2478,7 @@ fmountlens[] = {
         oss.copyfmt(os);
         char sign = value.toLong() < 0 ? '-' : '+';
         long h = long(abs(value.toFloat())/60.0);
-        long min = abs(value.toFloat()) - h*60;
+        long min = long(abs(value.toFloat()) - h*60);
         os << std::fixed << "UTC " << sign << std::setw(2) << std::setfill('0') << h << ":" 
            << std::setw(2) << std::setfill('0') << min;
         os.copyfmt(oss);
